Create TipSimpleCarousel component
